AB#334: Also rename the assert to AuthUtils.assertCanAdministrateOrg
And update the comments and logging a bit.

+++ src/main/java/mil/dds/anet/resources/PositionResource.java
@@ -53,7 +53,7 @@ public class PositionResource {
       final Position existingPos = dao.getByUuid(pos.getUuid());
       if (existingPos.getType() != PositionType.SUPER_USER) {
         throw new WebApplicationException(
-            "You are not allowed to change the position type to SUPER USER", Status.FORBIDDEN);
+            "You are not allowed to change the position type to SUPER_USER", Status.FORBIDDEN);
       }
     }
   }
@@ -66,7 +66,7 @@ public class PositionResource {
       throw new WebApplicationException("A Position must belong to an organization",
           Status.BAD_REQUEST);
     }
-    AuthUtils.assertSuperUserForOrg(user, pos.getOrganizationUuid(), true);
+    AuthUtils.assertCanAdministrateOrg(user, pos.getOrganizationUuid(), true);
   }
 
   @GraphQLMutation(name = "createPosition")
@@ -94,7 +94,7 @@ public class PositionResource {
   public Integer updateAssociatedPosition(@GraphQLRootContext Map<String, Object> context,
       @GraphQLArgument(name = "position") Position pos) {
     final Person user = DaoUtils.getUserFromContext(context);
-    AuthUtils.assertSuperUserForOrg(user, pos.getOrganizationUuid(), true);
+    AuthUtils.assertCanAdministrateOrg(user, pos.getOrganizationUuid(), true);
 
     final Position current = dao.getByUuid(pos.getUuid());
     if (current == null) {
@@ -205,7 +205,7 @@ public class PositionResource {
     if (pos == null) {
       throw new WebApplicationException("Position not found", Status.NOT_FOUND);
     }
-    AuthUtils.assertSuperUserForOrg(user, pos.getOrganizationUuid(), true);
+    AuthUtils.assertCanAdministrateOrg(user, pos.getOrganizationUuid(), true);
 
     final int numRows = dao.setPersonInPosition(DaoUtils.getUuid(person), positionUuid);
     AnetAuditLogger.log("Person {} put in Position {} by {}", person, pos, user);
@@ -220,7 +220,7 @@ public class PositionResource {
     if (pos == null) {
       throw new WebApplicationException("Position not found", Status.NOT_FOUND);
     }
-    AuthUtils.assertSuperUserForOrg(user, pos.getOrganizationUuid(), true);
+    AuthUtils.assertCanAdministrateOrg(user, pos.getOrganizationUuid(), true);
 
     final int numRows = dao.removePersonFromPosition(positionUuid);
     if (numRows == 0) {
